refactor: move list.go logic in clusterlogs.go
